requirements.txt
appdirs==1.4.3
attrs==19.3.0
Automat==0.8.0
beautifulsoup4==4.8.2
bokeh==1.4.0
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
fake-useragent==0.1.11
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.18.1
packaging==20.1
parse==1.14.0
parsel==1.5.2
Pillow==7.0.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
pyee==7.0.1
PyHamcrest==2.0.0
pymongo==3.10.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyppeteer==0.0.25
PyQt5==5.13.0
PyQt5-sip==12.7.1
PyQt5-stubs==5.13.1.4
pyqt5-tools==5.13.0.1.5
pyquery==1.4.1
python-dateutil==2.8.1
python-dotenv==0.11.0
PyYAML==5.3
queuelib==1.5.0
redis==3.4.1
requests==2.22.0
requests-html==0.10.0
Scrapy==1.8.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
soupsieve==1.9.5
tornado==6.0.3
tqdm==4.42.1
Twisted==19.10.0
urllib3==1.25.8
vthread==0.1.1
w3lib==1.21.0
websockets==8.1
zope.interface==4.7.1
